# openapi.yaml
openapi: 3.0.3
info:
  title: Recipes From Gramma and Others
  description: |
    This is a service for recipes for everyone.
  contact:
    email: botvimar@post.bgu.ac.il
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/MarinaBotvinnik/API_3.1/1.0.0
  description: SwaggerHub API Auto Mocking
- url: http://localhost:3000
  description: localhost
...
paths:
...
  /recipe/search/query/{searchQuery}:
    get:
      tags:
      - recipe
      summary: search a recipe
      description: choose options as you desire, and search for previews for recipes
      parameters:
      - name: searchQuery
        in: path
        description: search for recipes in a query given
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: numOfRecipes
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: intolerances
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: cuisine
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: diet
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: recipes with their preview details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/recipePreview'
        "404":
          description: No recipes found
...
  /recipe/recipeByID:
    get:
      tags:
      - recipe
      summary: get recipes by recipe name
      description: get recipes by recipe name and return properties about it
      parameters:
      - name: recipeid
        in: path
        description: The id of the recipe
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: recipe properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/recipePreview'
        "404":
          description: no recipes with this name
  /recipe/Information/{recipeID}:
    get:
      tags:
      - recipe
      summary: get recipe by information by ID
      description: get recipe by ID and return information about it
      parameters:
      - name: recipeID
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: the recipe preview information with extra details for preperation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/recipeInfo'
        "404":
          description: Recipe not found
  /recipe/addToFavorites:
    post:
      tags:
      - recipe
      summary: add recipe to user favorites
      description: get recipe by ID and adds it to the user favorites
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/favoriteRecip'
      responses:
        "200":
          description: Recipe added to favorites successfully
        "401":
          description: Unauthorized user
        "404":
          description: recipe not found
components:
  schemas:
    favoriteRecip:
      type: object
      properties:
        recipe_id:
          type: string
      example:
        recipe_id: 1232
...
    recipePreview:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        image:
          type: string
        timeToMake:
          type: integer
        popular:
          type: integer
        vegetarian:
          type: boolean
        vegan:
          type: boolean
        glutten:
          type: boolean
        watched:
          type: boolean
        saved:
          type: boolean
      example:
        id: 6
        title: Cake
        image: image
        timeToMake: 30
        popular: 5
        vegetarian: true
        vegan: false
        glutten: true
        watched: false
        saved: true
    recipeInfo:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        image:
          type: string
        timeToMake:
          type: integer
        popular:
          type: integer
        vegetarian:
          type: boolean
        vegan:
          type: boolean
        glutten:
          type: boolean
        watched:
          type: boolean
        saved:
          type: boolean
        ingredients:
          type: string
        instructions:
          type: string
        servings:
          type: integer
      example:
        id: 6
        title: Cake
        image: image
        timeToMake: 30
        popular: 5
        vegetarian: true
        vegan: false
        glutten: true
        watched: false
        saved: true
        ingredients: ingredients
        instructions: instructions
        servings: 4